backend/reset-password/requirements.txt
bcrypt==4.1.2
psycopg2-binary==2.9.9
